# docs/swagger.yaml
basePath: /gateway/v1
definitions:
  dto.BaseFriendRequest:
    properties:
      friendUid:
        example: av8FMdRdcb
        type: string
    required:
    - friendUid
    type: object
...
  dto.ConfirmFriendRequestRequest:
    properties:
      friendRequestId:
        example: 2
        type: integer
    required:
    - friendRequestId
    type: object
...
  dto.Friend:
    properties:
      createdAt:
        example: 1579098983
        type: integer
      friendAvatar:
        example: https://www.example.com/friendAvatar.png
        type: string
      friendName:
        example: friendName
        type: string
      friendUid:
        example: av8FMdRdcb
        type: string
      status:
        description: '0: friend, 1: stranger, 2: blacklist'
        example: 0
        type: integer
      uid:
        example: av8FMdRdcb
        type: string
      updatedAt:
        example: 1579098983
        type: integer
    type: object
  dto.FriendRequest:
    properties:
      createdAt:
        example: 1579098983
        type: integer
      friendAvatar:
        example: https://www.example.com/friendAvatar.png
        type: string
      friendName:
        example: friendName
        type: string
      friendUid:
        example: av8FMdRdcb
        type: string
      id:
        example: 1
        type: integer
      status:
        description: '0: pending, 1: accepted, 2: rejected'
        example: 0
        type: integer
      uid:
        example: av8FMdRdcb
        type: string
      updatedAt:
        example: 1579098983
        type: integer
    type: object
...
  response.Meta:
    properties:
      extra:
        additionalProperties:
          type: string
        type: object
      page:
        example: 1
        minimum: 1
        type: integer
      pageSize:
        example: 10
        minimum: 1
        type: integer
      total:
        type: integer
    type: object
  response.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
      meta:
        $ref: '#/definitions/response.Meta'
      reason:
        type: string
    type: object
info:
  contact:
    email: yusankurban@gmail.com
    name: Yusank
    url: https://yusank.space
  description: GoIM.Gateway 服务器接口文档
  license:
    name: MIT
    url: https://github.com/go-goim/core/blob/main/LICENSE
  termsOfService: http://go-goim.github.io/
  title: GoIM.Gateway Swagger
  version: "1.0"
paths:
  /discovery/discover:
    get:
      description: 获取推送服务器 IP
      parameters:
      - description: token
        in: query
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: errCode
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: unauthorized
          schema:
            type: "null"
      summary: 获取推送服务器
      tags:
      - discover
...
  /user/friend/accept:
    post:
      consumes:
      - application/json
      description: 接受好友请求
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      - description: body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.ConfirmFriendRequestRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: error
          schema:
            $ref: '#/definitions/response.Response'
      summary: 接受好友请求
      tags:
      - 好友
  /user/friend/add:
    post:
      consumes:
      - application/json
      description: 添加好友
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      - description: body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.BaseFriendRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.FriendRequest'
                  type: array
              type: object
        "400":
          description: error
          schema:
            $ref: '#/definitions/response.Response'
      summary: 添加好友
      tags:
      - 好友
  /user/friend/block:
    post:
      consumes:
      - application/json
      description: 屏蔽好友
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      - description: body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.BaseFriendRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: error
          schema:
            $ref: '#/definitions/response.Response'
      summary: 屏蔽好友
      tags:
      - 好友
  /user/friend/delete:
    post:
      consumes:
      - application/json
      description: 删除好友
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      - description: body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.BaseFriendRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: error
          schema:
            $ref: '#/definitions/response.Response'
      summary: 删除好友
      tags:
      - 好友
  /user/friend/list:
    get:
      description: 获取好友列表
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      - description: page
        in: query
        name: page
        type: integer
      - description: size
        in: query
        name: size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.Friend'
                  type: array
              type: object
        "400":
          description: err
          schema:
            $ref: '#/definitions/response.Response'
      summary: 获取好友列表
      tags:
      - 好友
  /user/friend/reject:
    post:
      consumes:
      - application/json
      description: 拒绝好友请求
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      - description: body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.ConfirmFriendRequestRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: error
          schema:
            $ref: '#/definitions/response.Response'
      summary: 拒绝好友请求
      tags:
      - 好友
  /user/friend/request/list:
    get:
      description: 获取好友请求列表
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      - description: page
        in: query
        name: page
        type: integer
      - description: size
        in: query
        name: size
        type: integer
      - description: status
        in: query
        name: status
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.FriendRequest'
                  type: array
              type: object
        "400":
          description: err
          schema:
            $ref: '#/definitions/response.Response'
      summary: 获取好友请求列表
      tags:
      - 好友
  /user/friend/unblock:
    post:
      consumes:
      - application/json
      description: 取消屏蔽好友
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      - description: body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.BaseFriendRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: error
          schema:
            $ref: '#/definitions/response.Response'
      summary: 取消屏蔽好友
      tags:
      - 好友
...
swagger: "2.0"